test(scan): move source_tree_with_compile_db fixture to conftest.py
Per the tests/ convention (fixtures live in conftest.py), relocate the shared
source-tree fixture out of test_cli_scan.py so other suites can reuse it.
Addresses CodeRabbit review on #438.
